`refactor` components de ui movidos para pasta ui
