+++ App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { NavigationContainer } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
-// theme
+// ui
 import { Colors } from './ui/Colors'
 
 // pages
