+++ pages/Operacoes.js
@@ -2,11 +2,9 @@ import React, { useState } from 'react'
 import { Text } from 'react-native'
 import { Picker } from '@react-native-picker/picker'
 
-// components
-import SafeArea from '../components/SafeArea'
-import Button from '../components/Button'
-
-// theme
+// ui
+import SafeArea from '../ui/SafeArea'
+import Button from '../ui/Button'
 import { Colors } from '../ui/Colors'
 
 export default function Operacoes({ navigation }) {
