+++ ui/Input.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { TextInput, View, StyleSheet } from 'react-native'
-import { Colors } from '../ui/Colors'
+import { Colors } from './Colors'
 
 export default function Input({
   value,
